Resolve issue #8
parse_smt2_string returns a AstVector class, which cannot be used further, e.g. with z3.AND. The first element of AstVector has to be used only. As we are parsing single asserts only, this behaviour should be save. However, it is unclear to me, why this has worked in the past.